Rewrite of user roles
