+++ lib/ansible/modules/storage/netapp/na_ontap_user_role.py
@@ -1,6 +1,6 @@
 #!/usr/bin/python
 
-# (c) 2018, NetApp, Inc
+# (c) 2018-2019, NetApp, Inc
 # GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 
 from __future__ import absolute_import, division, print_function
@@ -20,7 +20,7 @@ short_description: NetApp ONTAP user role configuration and management
 extends_documentation_fragment:
     - netapp.na_ontap
 version_added: '2.6'
-author: NetApp Ansible Team (@carchi8py) <ng-ansibleteam@netapp.com>
+author: NetApp Ansible Team (ng-ansibleteam@netapp.com)
 
 description:
 - Create or destroy user roles
@@ -62,7 +62,7 @@ EXAMPLES = """
       na_ontap_user_role:
         state: present
         name: ansibleRole
-        command_directory_name: DEFAULT
+        command_directory_name: volume
         access_level: none
         vserver: ansibleVServer
         hostname: "{{ netapp_hostname }}"
@@ -78,6 +78,7 @@ import traceback
 
 from ansible.module_utils.basic import AnsibleModule
 from ansible.module_utils._text import to_native
+from ansible.module_utils.netapp_module import NetAppModule
 import ansible.module_utils.netapp as netapp_utils
 
 
@@ -94,25 +95,20 @@ class NetAppOntapUserRole(object):
             command_directory_name=dict(required=True, type='str'),
             access_level=dict(required=False, type='str', default='all',
                               choices=['none', 'readonly', 'all']),
-            vserver=dict(required=True, type='str'),
+            vserver=dict(required=True, type='str')
         ))
 
         self.module = AnsibleModule(
             argument_spec=self.argument_spec,
             supports_check_mode=True
         )
-        parameters = self.module.params
-        # set up state variables
-        self.state = parameters['state']
-        self.name = parameters['name']
-        self.command_directory_name = parameters['command_directory_name']
-        self.access_level = parameters['access_level']
-        self.vserver = parameters['vserver']
+        self.na_helper = NetAppModule()
+        self.parameters = self.na_helper.set_parameters(self.module.params)
 
         if HAS_NETAPP_LIB is False:
             self.module.fail_json(msg="the python NetApp-Lib module is required")
         else:
-            self.server = netapp_utils.setup_na_ontap_zapi(module=self.module)
+            self.server = netapp_utils.setup_na_ontap_zapi(module=self.module, vserver=self.parameters['vserver'])
 
     def get_role(self):
         """
@@ -123,14 +119,14 @@ class NetAppOntapUserRole(object):
             False if role is not found
         :rtype: bool
         """
+        options = {'vserver': self.parameters['vserver'],
+                   'role-name': self.parameters['name'],
+                   'command-directory-name': self.parameters['command_directory_name']}
 
         security_login_role_get_iter = netapp_utils.zapi.NaElement(
             'security-login-role-get-iter')
         query_details = netapp_utils.zapi.NaElement.create_node_with_children(
-            'security-login-role-info', **{'vserver': self.vserver,
-                                           'role-name': self.name,
-                                           'command-directory-name':
-                                               self.command_directory_name})
+            'security-login-role-info', **options)
         query = netapp_utils.zapi.NaElement('query')
         query.add_child_elem(query_details)
         security_login_role_get_iter.add_child_elem(query)
@@ -141,68 +137,76 @@ class NetAppOntapUserRole(object):
         except netapp_utils.zapi.NaApiError as e:
             # Error 16031 denotes a role not being found.
             if to_native(e.code) == "16031":
-                return False
+                return None
             # Error 16039 denotes command directory not found.
             elif to_native(e.code) == "16039":
-                return False
+                return None
             else:
                 self.module.fail_json(msg='Error getting role %s: %s' % (self.name, to_native(e)),
                                       exception=traceback.format_exc())
         if (result.get_child_by_name('num-records') and
                 int(result.get_child_content('num-records')) >= 1):
-            return True
-        return False
+            role_info = result.get_child_by_name('attributes-list').get_child_by_name('security-login-role-info')
+            result = {
+                'name': role_info['role-name'],
+                'access_level': role_info['access-level'],
+                'command_directory_name': role_info['command-directory-name']
+            }
+            return result
+
+        return None
 
     def create_role(self):
-        role_create = netapp_utils.zapi.NaElement.create_node_with_children(
-            'security-login-role-create', **{'vserver': self.vserver,
-                                             'role-name': self.name,
-                                             'command-directory-name':
-                                                 self.command_directory_name,
-                                             'access-level':
-                                                 self.access_level})
+        options = {'vserver': self.parameters['vserver'],
+                   'role-name': self.parameters['name'],
+                   'command-directory-name': self.parameters['command_directory_name'],
+                   'access-level': self.parameters['access_level']}
+        role_create = netapp_utils.zapi.NaElement.create_node_with_children('security-login-role-create', **options)
+
         try:
             self.server.invoke_successfully(role_create,
                                             enable_tunneling=False)
         except netapp_utils.zapi.NaApiError as error:
-            self.module.fail_json(msg='Error creating role %s: %s' % (self.name, to_native(error)),
+            self.module.fail_json(msg='Error creating role %s: %s' % (self.parameters['name'], to_native(error)),
                                   exception=traceback.format_exc())
 
     def delete_role(self):
         role_delete = netapp_utils.zapi.NaElement.create_node_with_children(
-            'security-login-role-delete', **{'vserver': self.vserver,
-                                             'role-name': self.name,
+            'security-login-role-delete', **{'vserver': self.parameters['vserver'],
+                                             'role-name': self.parameters['name'],
                                              'command-directory-name':
-                                                 self.command_directory_name})
+                                                 self.parameters['command_directory_name']})
 
         try:
             self.server.invoke_successfully(role_delete,
                                             enable_tunneling=False)
         except netapp_utils.zapi.NaApiError as error:
-            self.module.fail_json(msg='Error removing role %s: %s' % (self.name, to_native(error)),
+            self.module.fail_json(msg='Error removing role %s: %s' % (self.parameters['name'], to_native(error)),
                                   exception=traceback.format_exc())
 
     def apply(self):
-        changed = False
-        vserver = netapp_utils.setup_na_ontap_zapi(module=self.module, vserver=self.vserver)
-        netapp_utils.ems_log_event("na_ontap_user_role", vserver)
-        role_exists = self.get_role()
-        if role_exists:
-            if self.state == 'absent':
-                changed = True
-        else:
-            if self.state == 'present':
-                changed = True
-        if changed:
+        self.asup_log_for_cserver('na_ontap_user_role')
+        current = self.get_role()
+        cd_action = self.na_helper.get_cd_action(current, self.parameters)
+
+        if self.na_helper.changed:
             if self.module.check_mode:
                 pass
             else:
-                if self.state == 'present':
-                    if not role_exists:
-                        self.create_role()
-                elif self.state == 'absent':
+                if cd_action == 'create':
+                    self.create_role()
+                elif cd_action == 'delete':
                     self.delete_role()
-        self.module.exit_json(changed=changed)
+        self.module.exit_json(changed=self.na_helper.changed)
+
+    def asup_log_for_cserver(self, event_name):
+        """
+        Fetch admin vserver for the given cluster
+        Create and Autosupport log event with the given module name
+        :param event_name: Name of the event log
+        :return: None
+        """
+        netapp_utils.ems_log_event(event_name, self.server)
 
 
 def main():
